Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.10.6
python-dotenv==1.0.0
python-socketio==5.10.0
bcrypt==4.1.2
pytz==2024.1
face-recognition==1.3.0
opencv-python==4.10.0.84
numpy==1.26.4
dlib==19.24.99